Avoid implicit auto-boxing

+++ src/test/java/net/objecthunter/exp4j/ConcurrencyTests.java
@@ -17,9 +17,6 @@ package net.objecthunter.exp4j;
 
 import org.junit.Test;
 
-import java.text.DateFormat;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
@@ -44,7 +41,7 @@ public class ConcurrencyTests {
                     .variables("pi", "n")
                     .build()
                     .setVariable("pi",Math.PI)
-                    .setVariable("n", i)
+                    .setVariable("n", (double) i)
                     .evaluateAsync(exec);
 
             correct2[i] = Math.log(Math.E * Math.PI * (i+1));
@@ -53,7 +50,7 @@ public class ConcurrencyTests {
                     .build()
                     .setVariable("pi",Math.PI)
                     .setVariable("e", Math.E)
-                    .setVariable("n", i)
+                    .setVariable("n", (double) i)
                     .evaluateAsync(exec);
         }
